refactor: move merging logic into reporter file
